pyproject.toml
[project]
name = "vlunar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "gymnasium",
    "gymnasium[box2d]",
    "moviepy",
    "gymnasium[other]",
    "swig",
    "torch",
    "transformers",
    "accelerate",
    "pillow",
    "torchvision",
    "peft",
    "bitsandbytes",
]
